requirements.txt
pandas
requests
pyyaml
lxml
yfinance
tensorflow
keras
scikit-learn
matplotlib
joblib
numpy
scipy
seaborn
